refactor: auth 폴더 리펙토링
