+++ src/auth/auth.service.ts
@@ -1,10 +1,10 @@
 import { BadRequestException, Injectable, InternalServerErrorException } from '@nestjs/common';
 import { PrismaService } from 'src/common/prisma/prisma.service';
 import * as bcrypt from 'bcrypt';
-import { SigninDto } from './dtos/signin.dto';
-import { SignupDto } from './dtos/signup.dto';
+import { SigninDto } from './dto/signin.dto';
+import { SignupDto } from './dto/signup.dto';
 import { JwtService } from '@nestjs/jwt';
-import { Payload } from './interfaces/payload.interface';
+import { Payload } from './interface/payload.interface';
 
 const salt = 10;
 
@@ -15,29 +15,47 @@ export class AuthService {
         private readonly jwtService: JwtService
     ) {}
 
-    async signup(signupData: SignupDto): Promise<void> {
-        if (await this.checkUsernameDuplicate(signupData.username)) {
+    async signup(dto: SignupDto): Promise<void> {
+        if (await this.checkUsernameDuplicate(dto.username)) {
             throw new BadRequestException("이미 사용중인 이름입니다");
         }
-        if (await this.checkEmailDuplicate(signupData.email)) {
+        if (await this.checkEmailDuplicate(dto.email)) {
             throw new BadRequestException("이미 사용중인 이메일입니다");
         }
-        try {
-            const password = await bcrypt.hash(signupData.password, salt)
-            await this.prismaService.users.create({ 
-                data : {
-                    username : signupData.username,
-                    password : password,
-                    email : signupData.email
-                }
-            });
-        } catch (err) {
-            console.log(err)
-            throw new InternalServerErrorException();
+
+        const password = await bcrypt.hash(dto.password, salt)
+        await this.prismaService.users.create({ 
+            data : {
+                username : dto.username,
+                password : password,
+                email : dto.email
+            }
+        });
+    }
+
+    async signin(dto: SigninDto): Promise<unknown> {
+        let findUser = await this.prismaService.users.findUnique({
+            where : { username : dto.username },
+        })
+
+        if (findUser) {
+            // @TODO 비밀번호 검사 안하는 중 (테스트 용)
+            // const match = await bcrypt.compare(signinData.password, findUser.password)
+            const match = true;
+            if (match) {
+                const payload: Payload = { userId: findUser.id, username: dto.username }
+                const jwt = { accessToken : this.jwtService.sign(payload) };
+                return jwt
+            } else {
+                throw new BadRequestException("아이디 또는 비밀번호가 잘못되었습니다");
+            }
+        } else {
+            throw new BadRequestException("아이디 또는 비밀번호가 잘못되었습니다,");
         }
     }
 
-    async checkUsernameDuplicate(username: string): Promise<Boolean> {
+    // 유저 이름 중복 체크
+    private async checkUsernameDuplicate(username: string): Promise<Boolean> {
         try {
             const result = await this.prismaService.users.findUnique({
                 where : {  username : username }
@@ -49,7 +67,8 @@ export class AuthService {
         }
     }
 
-    async checkEmailDuplicate(email: string): Promise<Boolean> {
+    // 이메일 중복 체크
+    private async checkEmailDuplicate(email: string): Promise<Boolean> {
         try {
             const result = await this.prismaService.users.findUnique({
                 where : { email : email }
@@ -58,32 +77,6 @@ export class AuthService {
         } catch (err) {
             console.log(err)
             return false;
-        }
-    }
-
-    async signin(signinData: SigninDto): Promise<unknown> {
-        let findUser;
-        try {
-            findUser = await this.prismaService.users.findUnique({
-                where : { username : signinData.username },
-            })
-        } catch (err) {
-            console.log(err)
-            throw new InternalServerErrorException("서버에 오류가 발생했습니다")
-        }
-
-        if (findUser) {
-            // const match = await bcrypt.compare(signinData.password, findUser.password)
-            const match = true;
-            if (match) {
-                const payload: Payload = { userId: findUser.id, username: signinData.username }
-                const jwt = { accessToken : this.jwtService.sign(payload) };
-                return jwt
-            } else {
-                throw new BadRequestException("아이디 또는 비밀번호가 잘못되었습니다");
-            }
-        } else {
-            throw new BadRequestException("아이디 또는 비밀번호가 잘못되었습니다,");
         }
     }
 }
