+++ src/stocks/account/account.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Post, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
-import { Payload } from 'src/auth/interfaces/payload.interface';
+import { Payload } from 'src/auth/interface/payload.interface';
 import { GetUser } from 'src/common/decorator/get-user.decorator';
 import { AccountService } from './account.service';
